requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
mysql-connector-python==9.1.0
chromadb==0.5.18
google-generativeai==0.8.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
reportlab==4.2.2
python-docx==1.1.2

